11th new branch may be many done

+++ src/entities/MobileAccount.java
@@ -17,15 +17,12 @@ public class MobileAccount extends Account implements Serializable{
     private int creditCardNumber;
     private int debitCardNumber;
     private int subscriberId;
-    
 
     public MobileAccount(int subscriberId,Date createdDate) {
         super(createdDate);
         this.subscriberId = subscriberId;
     }
-
-    
-    public MobileAccount(String accountNumber, double amount, Date createdDate, Date expireDate, String packageType,int payPalAccount, int creditCardNumber,int debitCardNumber,int subscriberId) {
+    public MobileAccount(String accountNumber, float amount, Date createdDate, Date expireDate, String packageType,int payPalAccount, int creditCardNumber,int debitCardNumber,int subscriberId) {
         super(accountNumber, amount, createdDate, expireDate, packageType);
         
         this.payPalAccount=payPalAccount;
@@ -67,7 +64,7 @@ public class MobileAccount extends Account implements Serializable{
         this.subscriberId = subscriberId;
     }
     
-   public void topUp(double amount) {
+   public void topUp(float amount) {
         super.setAmount(super.getAmount() + amount);
 
     }
@@ -77,7 +74,7 @@ public class MobileAccount extends Account implements Serializable{
 
     }
 
-    public void deductAmount(double amount) {
+    public void deductAmount(float amount) {
         super.setAmount(super.getAmount() - amount);
 
     }
